ramlat: implement memory latency measurements
This program measures the memory latency, or more exactly the transaction
rate on random addresses per word size and per area size.

The principle consists in walking over the memory using random addresses
and to measure how many accesses are possible per unit of time. The base
addresses are all multiples of 64 bytes so that it is possible to measure
the time taken to retrieve 1 to 64 bytes without being badly affected by
random cache line crossing.

An initial version was imperfect because it was using a fast xorshift
random pattern generator which used to consume CPU cycles and registers,
hence used to produce much lower numbers than expected, especially in
L1/L2 caches.

This new version drops the xorshift in favor of a modular arithmetic
approach, consisting in adding a constant prime value to the existing
address. In order to limit the number of operations, a fixed prime is
chosen (257), multiplied by 512, and accesses within the 512-byte range
are hard-coded using simple offsets that most CPUs can trivially add to
a pointer. This results in a random enough pattern that prevents any
form of prefetching from being performed by the cache controller and
which emphasizes the differences between in-cache and out-of-cache
performance.

The values are read by the compiler as the argument to an empty "asm"
statement that doesn't use them. It ensures they are not optimized away
and that the compiler does its best to fetch the data. Whenever possible
we try to read multiple words in the same asm statement. That avoids the
write-after-write dependency which happens when the compiler wants to
read everything using the same register because it doesn't know that we
didn't use the value previously fetched.

On ix86 (32-bit), there are too few registers and the performance suffers
when multiple reads are passed to the same asm statement because the last
read touches a register that is also used somewhere else in the loop. For
this reason we check the architecture to either perform two consecutive
reads or a double one.

It is worth mentionning that on all archs tested (i586, x86_64, mips,
armv7t, armv8), performing a 64-bit read at once is much more efficient
than implementing it as two 32-bit reads, apparently because the compiler
knows that it can always use a certain pair of registers and produces
better code this way. So now instead of deciding whether we read 64 bits
at once based on the word size, we only decide based on the risk of
register shortage (ie ix86) whether we read a single 64-bit integer or
two at once.