ramlat: replace the bogus walk algorithm
A bug was detected on a core i5 3320M, where the performance in the
L1 cache appears twice as fast as reality and where the performance
slowly degrades as the size grows. The cause was identified as being
the addition of the mask to addr1 to form addr2, because the mask in
fact has lots of lower bits set and doesn't cause the next addresses
to be aligned.

It turns out that the L1 cache in the CPU above seems to have two
ports which are selected based on bits 3..5 of the address. If two
addresses are aligned modulo 64, the same port is used and a single
access is possible. If they differ by at least 8 bytes modulo 64,
then two concurrent accesses are possible. Here the mask was always
causing such a difference to appear.

The slow performance degradation when the area size increases is
caused by the fact that the mask overlaps with some of the next
addresses to be fetched, thus a significant percentage of the read
operations from L2 cache are in fact found in the L1 cache.

And since it's important to preserve registers during these accesses,
it's worth replacing the xor with an addition so that the compiler
may easily emit constant offsets around a single register.

During the debugging, it was also noticed that the 1kB test was already
bogus because due to the way the xor is performed, accesses outside of
the 1kB area were already made.

Last, it was observed that some slower in-order CPUs still waste a lot
of time processing operations instead of accessing memory.

This was enough to redesign the walk algorithm to perform the same
pseudo-random walk over 8 areas of 512 bytes each, split in a way that
helps the compiler group accesses as much as possible without wasting
registers nor operations.

With this fix, the performance now shows that the core i5 above performs
one cache read per CPU cycle, a core i7 6700K performs two cache reads
per CPU cycle (suggesting the L1 cache doesn't conflict anymore between
similar addresses), the AR9331 and the Armada370 do 1 read every 2 CPU
cycles in L1.