ramlat: use direct pointers instead of base pointer and offsets
On ARMv7 the L1 cache performance looked suspiciously off by a few
percent compared to the expected cache frequency. Upon inspection
of the code, it appeared that the compiler sometimes had a tough
job at updating the relative offsets, and used to have to stuff
up to 3 "add" operations between two "ldrd". This would also change
with the compiler version, proving that we were still measuring the
CPU's performance and compiler's ability to optimize the code.

On x86 there's no such problem because we support a complex addressing
with base + offset + immediate and the immediate may be as large as
desired.

So let's change the code to use pointers only. This will still require
operations as the compiler may now simply update the pointers and a
relative offset for each read.

The performance has now significantly improved on ARMv7, to match the
CPU's frequency on 32-bit reads, and on MIPS it has even doubled for
the same reason. No change was observed on x86 as expected.