scripts: add scripts to produce some bogolocs and to graph them
BogoLocs means "bogus lines of code per second". They're a non-scientific
measurement of memory performance of a device across the L1/L2 caches and
main memory that returns comparable values to what is observed by compiling
code. They are just useful to have a metric to observe a device's behaviour
under different workload.

Since a compiler spends a lot of time manipulating lists and trees, here
the approach consists in only looking at the word size that corresponds
to twice a pointer size (one access for the pointer, another one for the
data). We use a 96% L1 hit ratio and 92% L2 hit ratio to get a more or
less valid result. The L2 hit rate is slightly decreased as the number
of cores increase, to take into account the fact that it is shared on
most devices.