ramlat: change the read calls to split address and offset
Some instructions support an immediate offset but if the instructions have
to be cut into several smaller instructions, then there's a loss of efficiency
because we're forced to perform some local computation for the additional
offsets. Let's change the way everything works in order to have only base
and offsets. That way it will be possible to change the offset only
directly from the asm code.