ramlat: remove the "loop" variable from the loops
Since we're adding a prime number, we can simply check that the final value
is equal to LOOPS_PER_ROUND * prime * mask, so we don't need the loop index.
In order to build it easier we make the loop start from the end, it simplifies
the tests. In addition to saving one possible register in the fast path, it
has slightly improved the performance on the 128-bit access using VFP (+1%).