ramlat: implement dual-128 and dual-256
Both of them are more efficient with VFP, allowing to recover the
base transfer rate. The code needs to be refactored, the difference
in offset is always 512, we can hard-code this. Also we're still
wasting time adding integers, the code needs to be rewritten with
a specific loop for each word size and possibly per architecture.