ramlat: massive code reorganisation in order to loop in functions
Instead of looping in the main function, we now have individual run*
functions for each word size. These functions are registered in an
array indexed by word size so that it will even be possible to use
different functions when implementations permit.

For now the functions have all been called "generic". They still
rely on the existing read* functions.