ramlat: vfp: reduce FP write after write dependency on 64-dual
On Armada370, the VFP performance in the L1 cache is lower than V7's.
The reason is that we're constantly using d4 and d5 only. By changing
the read function to process 4 64-bit words at once instead, using
d4,d5,d6,d7, the performance is now the same as armv7. Cortex A17 was
not affected here.