ramlat: fix ARMv7's read64_dual to force the target registers
Gcc can assign a random pair of registers, to the target registers
of ldrd, even a combination causing the program to crash. Anyway
that prevents from building in ARM mode because registers must be
adjacent, even and odd. Let's fix them to r0 and r1 only.

We must also avoid passing an offset > 255 in the address field
because it's limited to 8 bit in ARM mode. Thus we now compute the
offset by hand, and we simplify the call so that 512 is systematically
added by the read function.