ramlat: reduce the number of operations on read64_armv7
The problem in ARM mode is that the offsets being limited to 255,
we have as many additions as register loads, effectively limiting
to one load per cycle on a dual-issue CPU minus the overhead of
the loop.

One reason is the way our offsets are chosen because it's impossible
to assign registers with certain offsets in order to reuse them. But
we don't want to produce some contiguous loads to avoid the risk of
streaming. So let's slightly change the walk pattern in order to
allow two register to progress 512 apart within a 1024 block (as is
currently the case), but change the second offset by 128 immediately
so that the registers can be reused. The offsets now follow the
equivalent of 4 embedded loops :

    1) 64
    2) 256
    3) 128
    4) 512

The addresses are then visited in this order :

      0, 512, 128, 640, 256, 768, 384, 896,
     64, 588, 192, 704, 320, 832, 448, 960

The read function was changed to read quads so that it performs the
+128 addition in the LDRD offset. And we can successfully have 4
loads for 2 additions. Now it's possible to saturate the cache with
64 bit words even in ARM mode.