ramlat: add an ARMv7 version of run128_dual.
This one uses ldrd to read 2 64-bit operands twice, and to do it again
for the second offset. The performance is the same as in generic mode
since the compiler also emits such instructions.

On the Armada 370, the performance becomes better if the second load of
the first word is swapped with the first load of the second word. However
that defeats the purpose of the test since the test consists in checking
how fast we can process independant words, so they were left in the
expected order.