ramlat: reduce the loop overhead
Compared to earlier numbers, some numbers were slightly off, typically
by 1/8 to 1/32. In fact the loop's cost had increased on mips and ARM
because the operation on rnd requires two instructions, so the test
performed just after has to wait for these two instructions to complete.

For this reason now the loop only checks for the end and decreases rnd
later. The compiler notices that rnd is independant on the first load
and manages to interleave it with the first two load instructions (when
the code is not in asm). For the asm code, we simply put the decrement
manually after the first call. Depending on the asm code complexity it
is efficient or not.

All tests now show slightly better numbers. The loop cost seems to be
around 5 cycles, maybe it's possible to do slightly better.