ramlat: better optimize the armv8 256-bit reads
The compiler cannot place the pointer increment optimally, resulting
in 6% lower perf (17 cycles needed for 16 transfers). This open-codes
the 256-bit read of a 4kB block and automatically increments the
pointer. The performance is much better and could possibly further
be improved. The same method should be implemented for other sizes.