ramlat: implement indexed and direct memory accesses for 32, 64, void*
The previous memory accesses wouldn't correctly measure the latency because
the processor was able to program prefetches based on the instructions in
the queue. This was particularly visible on x86 where DRAM accesses were
much too fast.

Now we fill the memory area with a butterfly pattern made of inverted bits
which are the worst case for a cache, and jump between memory locations.
This guarantees that it's not possible for the system to prefetch memory
faster than it is available and that we really measure back-to-back
operations.

This mechanism already shows that depending how the accesses are performed
(direct pointer dereference or indexed location), the access time can differ
by one cycle on x86. For this reason we provide native pointers, 32 and 64
bit indexed locations. This directly shows how programs will be impacted
depending on the way they use memory.

We use with 1, 2 and 4 parallel pattern walks to observe the effect of the
number of memory channels.

Typical measures now indicate on x86_64 :
  - 4 cycles for L1 direct
  - 5 cycles for L1 indexed
  - 12 cycles for L2 direct
  - 13 cycles for L2 indirect
  - 65 ns for DRAM

The performance is now reported in accesses per millisecond instead of
accesses per microsecond, because the DRAM is far too slow to report
useful numbers otherwise.