ramlat: rename to rambw
The ramlat term was not suitable given that we measure throughput instead
of real latency when the memory controller is able to detect pending read
requests and merge them. However it contains various methods to optimize
the read rate so it still makes sense to detect memory bus bottlenecks on
this basis.

For this reason it was renamed to rambw since it measures bandwidth. It
should be adapted to measure only bandwidth (-b) and to stop trying to
perform a complex walk pattern.