ramlat: add support for configuring the work sizes
Instead of running all tests from 4kB to the max, it's now possible to
pass the list of power-of-two work sizes after -w, and in this case only
these ones will be tested. If no size is forced on the command line, the
largest one from the work set is automatically used.

For example it's now possible to limit the test execution time when running
in L1, L2, L3, RAM:

  $ time ./ramlat -s -c 1,3,5 -w 13,17,21,25 -n 200
     size:  1x32  1x64 1xPTR
       8k: 1.203 1.427 0.961
     128k: 2.899 3.121 2.883
    2048k: 11.55 11.69 11.44
   32768k: 69.28 66.97 65.48

  real    0m3.006s
  user    0m3.002s
  sys     0m0.004s