ramlat: improve the work area filling algorithm
The area is filled following a butterfly pattern, but currently
this is extremely slow because the filling follows the pattern. It's
particularly visible on large work areas which can take several
seconds to fill on small machines since each write of a single value
causes a cache miss that forces to commit one line and fetch a
complete one to change only one word.

The algorithm is not complicated to linearize as is easily shown when
printing all offsets, it requires to start from half of the size and
going up until the previous value (or size) is reached, after which we
must start from half of the previous initial value. Since we're using
powers of two it means stopping at a power of two and dividing the new
value by 4, until the end is reached.

With this change, tests are up to 8 times faster on small machines.