data: add results for armada-xp on the XP-GP board
The results are quite stable, and allow to detect when the board is
running with L2 cache disabled (this seems to happen as the result
of entering cpuidle state 2 which powers it down and doesn't seem
to power it up when leaving the state). Sizes between 64k and 2M
are around 20ns when enabled, and 88ns when disabled.