ramwalk: implement a simple ram walk pattern
This performs a pseudo-random RAM walk pattern over 1GB of RAM. In
fact it really is only a butterfly walk but that's sufficient to
detect when access times can cause performance degradations to
memory-heavy applications such as compilers. Execution times of +80%
were encountered on rk3588 between DDR4 and DDR5 when using DDR init
code v1.16.

It takes an optional number of rounds in argument to amortize the
initialization time. It defaults to 1. This just needs to be called
under "time".

Measures:
  i7-6700k, DDR4-2400 2x64 bits in 4 DIMMs, 10 rounds: 176.7s
  Altra Q80-26, DDR4-2933, 6x64 bits in 6 DIMMs, 10 rounds: 465.1s
  rk3588,   DDR4-4224 4x16 bits in 2 chips, 10 rounds: 417.4s
  rk3588,   DDR5-4800 4x16 bits in 2 chips, 10 rounds: 745.6s