rambw: only perform linear accesses in 512-bit mode and measure BW
The original rambw utility was in the middle between measuring latency
and bandwidth, because it was focusing on measuring the bandwidth of
random walks, which is not very relevant and is already achieved by
ramlat. There's one situation where it could make sense, which is to
check the effect of multiple memory controllers, but it wasn't even
using multiple threads, which restricts the ability to access multiple
locations to the length of the tested core's data prefetch queue.

Let's get back to something cleaner: we now only use 512-bit loads to
benefit from fully loaded cache lines, and we perform linear accesses,
that are supposed to be optimized and predicted by the memory controller
which tries to sustain a high bandwidth.