rambw: start a configurable number of threads
Now when passing -t we start as many thread minus one. The first thread
remains in foreground for simplicity. Any thread can wake up for the
signal handler, and will update the counters and the timer. If in the
future it would cause trouble with the time resolution, we could pin
that to a specific thread.

The measure is incorrect now because all threads start at the same place
hence each of them benefits from the other one's cache. We need to keep
them apart. If that turns out unreliable (e.g. some catching up with
others), then we may end up with as many areas.