rambw: assign a specific area to each thread
Now it works much better. The memset() is done in the thread itself
(better for CPU locality on startup). The wait time should be moved to
the threads as well.

However the sizing should change to be global, because now it's
complicated for the user to assign a correct size without risking to
consume too much RAM. Ideally we should specify the total size of the
area and the program would split it per thread (and possibly round
down to the nearest power of 2).