rambw: move the slowstart phase to the threads
Instead of doing the slowstart in the main thread and then starting
all of them, we're now making each thread loop until slowstart is zero,
and the timer handler will just zero that value when it kicks in. This
way all threads are ready to run at full speed ASAP, and if any
frequency limitation is enforced, it will already be applied.