rambw: add an estimate of the DDR-rating based on the bandwidth
Most boards show DRAM efficiency around 60, so by presetting an
efficiency ration to such a value, and specifying a bus width,
it becomes possible to suggest a DDR rating for the memory being
tested. Of course it remains very approximative given that the
efficiency really varies from one CPU to another, but it is
sometimes easier to compare than raw numbers.

It's worth noting that older CPUs with FSB and northbridges have a
higher efficiency based on this FSB and a lower one based on the
DRAM modules. Some more modern x86 machines show efficiencies up
to 85-88% for interleaved dual-rank chips, while others used in
servers can show around 35%. Also some devices seem to be using
their two channels one at a time (e.g. N5105) given that the
reported speed corresponds to half of the expected performance
level.

The main purpose is rather to compare boards between each other,
and possibly to spot an incorrect setting when the reported numbers
may hint about where the problem is (e.g; single vs dual-channel),
than publishing numbers.

When the efficiency is not set, it defaults to 60%.